requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.3
django-debug-toolbar==3.2.4
django-filter==21.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
